Atualização da arquitetura e adição do uso de hook no componente filtroselectsearch
