+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes} from 'react-router-dom';
 import HomePage from './HomePage';
-import TestPageRef from './TestPageRef';
+import TestePage from './TestePage';
 import './App.css';
 
 function App() {
@@ -9,7 +9,7 @@ function App() {
     <Router>
       <Routes>
         <Route path="/" element={<HomePage />} />
-        <Route path="/test-ref" element={<TestPageRef />} />
+        <Route path="/test-ref" element={<TestePage />} />
       </Routes>
     </Router>
   );
