--- src/components/styles.css
@@ -1,44 +1,0 @@
-/* styles.css */
-.container {
-    display: flex;
-    justify-content: space-between;
-    gap: 10px;
-    padding: 10px;
-    max-width: 1000px;
-  }
-  
-  .text-field-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .text-field-container label {
-    margin-bottom: 5px;
-    font-weight: bold;
-    color: white; /* Ajuste conforme necessário */
-  }
-  
-  .input-container {
-    position: relative;
-    width: 200px; /* Ajuste conforme necessário */
-  }
-  
-  .text-field {
-    padding: 10px;
-    padding-left: 30px; /* Espaço para o ícone */
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    font-size: 16px;
-    width: 100%;
-  }
-  
-  .icon {
-    position: absolute;
-    left: 10px;
-    top: 50%;
-    transform: translateY(-50%);
-    pointer-events: none;
-    color: #aaa; /* Ajuste conforme necessário */
-  }
-
